ROV-GUI/ROV-GUI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2022.3.29=haa95532_0
certifi=2021.10.8=py38haa95532_2
imageio=2.16.2=pypi_0
numpy=1.22.3=pypi_0
opencv-python=4.5.5.64=pypi_0
openssl=1.1.1n=h2bbff1b_0
pillow=9.1.0=pypi_0
pip=21.2.2=py38haa95532_0
python=3.8.13=h6244533_0
setuptools=61.2.0=py38haa95532_0
sqlite=3.38.2=h2bbff1b_0
tk=0.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
